Polish apply_to_static
